backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
slowapi==0.1.9
PyJWT>=2.8.0,<3.0
httpx>=0.27.0,<1.0
python-multipart>=0.0.9
libsql-experimental>=0.0.34
stripe>=8.0.0,<10.0
python-dotenv>=1.0.0
